+++ lib/src/dart3_suggestors/required_props/collect/aggregated_data.sg.dart
