+++ lib/src/dart3_suggestors/required_props/collect/analysis.dart
