+++ lib/src/dart3_suggestors/required_props/collect/collected_data.sg.dart
