+++ lib/src/dart3_suggestors/required_props/collect/package/metadata.dart
