+++ lib/src/dart3_suggestors/required_props/collect/util.dart
